+++ src/plugins/error-messages.js
@@ -144,8 +144,9 @@ const CODES = {
     SC19: 'SchemaCheck: compoundIndexes must contain arrays',
     SC20: 'SchemaCheck: compoundIndexes.array must contains strings',
     SC21: 'SchemaCheck: given index is not defined in schema',
-    SC22: 'given indexKey is not type:string',
+    SC22: 'SchemaCheck: given indexKey is not type:string',
     SC23: 'SchemaCheck: fieldname is not allowed',
+    SC24: 'SchemaCheck: required fields must be set via array. See https://spacetelescope.github.io/understanding-json-schema/reference/object.html#required',
 
     // plugins/validate.js
     VD1: 'Sub-schema not found, does the schemaPath exists in your schema?',