+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Breaking:
   - Upgraded to [pouchdb 7.0.0](https://pouchdb.com/2018/06/21/pouchdb-7.0.0.html)
   - `disableKeyCompression` is set to `false` by default
   - `RxDatabase.collection()` now only accepts the json-schema as schema-attribute
+  - It is no longer allowed to set required fields via `required: true`, use `required: ['myfield']` in compliance with the jsonschema standard
 
 Other:
   - Upgrade to eslint 5 (no more babel-eslint)